- added swagger - deleted unused hello and helloGet functions
