+++ product-service/package.json
@@ -5,6 +5,7 @@
   "main": "serverless.ts",
   "scripts": {
     "test": "jest",
+    "swagger": "sls generate-swagger",
     "deploy:dev": "sls deploy --stage dev",
     "deploy:getProductsList": "sls deploy function -f getProductsList",
     "deploy:getProductsById": "sls deploy function -f getProductsById",
@@ -27,6 +28,7 @@
     "jest": "^29.4.3",
     "json-schema-to-ts": "^1.5.0",
     "serverless": "^3.0.0",
+    "serverless-auto-swagger": "^2.12.0",
     "serverless-esbuild": "^1.23.3",
     "ts-jest": "^29.0.5",
     "ts-node": "^10.4.0",
