+++ product-service/src/domain/typings.ts
@@ -1,0 +1,10 @@
+export type ProductId = string;
+
+export interface Product {
+    id: ProductId,
+    title: string,
+    description: string,
+    price: number,
+    img: string,
+    count: number
+}
