--- product-service/src/functions/hello/handler.ts
@@ -1,16 +1,0 @@
-import type { ValidatedEventAPIGatewayProxyEvent } from '@libs/api-gateway';
-import { formatJSONResponse } from '@libs/api-gateway';
-import { middyfy } from '@libs/lambda';
-
-import schema from './schema';
-
-const hello: ValidatedEventAPIGatewayProxyEvent<typeof schema> = async (event) => {
-  const { name } = JSON.parse(event.body.toString());
-
-  return formatJSONResponse({
-    message: `Hello ${name}, welcome to the exciting Serverless world!`,
-    event,
-  });
-};
-
-export const main = middyfy(hello);
