+++ product-service/src/functions/index.ts
@@ -1,4 +1,2 @@
-export * from './hello';
-export * from './helloGet';
 export * from './getProductsList';
 export * from './getProductById';
